I updated namespaces, project name, assembly name and packages as per sample-sdk.
