--- Relativity.Transfer.Sample/App.config
@@ -1,26 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<configuration>
-    <appSettings>
-        <!--TransferMode acceptable values are: Aspera, Fileshare.-->
-        <add key="TransferMode" value="Aspera"/>
-        <add key="RelativityUrl" value="https://relativity_host.com"/>
-        <add key="RelativityUserName" value="jsmith@example.com"/>
-        <add key="RelativityPassword" value="UnbreakableP@ssword777"/>
-        <add key="WorkspaceId" value="123456"/>
-    </appSettings>
-    <startup>
-        <supportedRuntime version="v4.0" sku=".NETFramework,Version=v4.6.2"/>
-    </startup>
-    <runtime>
-        <assemblyBinding xmlns="urn:schemas-microsoft-com:asm.v1">
-            <dependentAssembly>
-                <assemblyIdentity name="Newtonsoft.Json" publicKeyToken="30ad4fe6b2a6aeed" culture="neutral"/>
-                <bindingRedirect oldVersion="0.0.0.0-13.0.0.0" newVersion="13.0.0.0"/>
-            </dependentAssembly>
-            <dependentAssembly>
-                <assemblyIdentity name="Renci.SshNet" publicKeyToken="1cee9f8bde3db106" culture="neutral"/>
-                <bindingRedirect oldVersion="0.0.0.0-2020.0.2.0" newVersion="2020.0.2.0"/>
-            </dependentAssembly>
-        </assemblyBinding>
-    </runtime>
-</configuration>
