+++ Sample.sln
@@ -1,9 +1,9 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 14
-VisualStudioVersion = 14.0.25420.1
+# Visual Studio Version 17
+VisualStudioVersion = 17.5.33502.453
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Relativity.Transfer.Sample", "Relativity.Transfer.Sample\Relativity.Transfer.Sample.csproj", "{6909F5D1-18F2-4A5A-B32B-D775D73F4647}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Relativity.Server.Transfer.SDK.Samples", "Server.Transfer.SDK.Samples\Relativity.Server.Transfer.SDK.Samples.csproj", "{6909F5D1-18F2-4A5A-B32B-D775D73F4647}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{3F5C1A02-919B-4FF9-B764-D5D5226F97BD}"
 	ProjectSection(SolutionItems) = preProject
