+++ Server.Transfer.SDK.Samples/LogConfig.xml
