+++ Server.Transfer.SDK.Samples/Resources/EDRM-Sample2.pdf
