+++ Server.Transfer.SDK.Samples/Resources/EDRM-Sample3.doc
