+++ Server.Transfer.SDK.Samples/nuget.config
