+++ Server.Transfer.SDK.Samples/packages.config
@@ -1,0 +1,15 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="Newtonsoft.Json" version="13.0.2" targetFramework="net462" />
+  <package id="Polly" version="5.3.1" targetFramework="net462" />
+  <package id="Relativity.Server.DataTransfer.SDK" version="0.9.0" targetFramework="net462" />
+  <package id="Relativity.Server.Kepler.Client.SDK" version="2.15.6" targetFramework="net462" />
+  <package id="Relativity.Server.Kepler.SDK" version="2.15.6" targetFramework="net462" />
+  <package id="Relativity.Server.Logging.Interfaces.SDK" version="3000.4.4" targetFramework="net462" />
+  <package id="Relativity.Server.Logging.SDK" version="3000.4.4" targetFramework="net462" />
+  <package id="Relativity.Server.Services.DataContracts.SDK" version="13.6.1" targetFramework="net462" />
+  <package id="Relativity.Server.Services.Interfaces.SDK" version="13.6.1" targetFramework="net462" />
+  <package id="Relativity.Server.Transfer.SDK" version="7.7.0" targetFramework="net462" />
+  <package id="SSH.NET" version="2016.0.0" targetFramework="net462" />
+  <package id="System.Reactive" version="4.0.0" targetFramework="net462" />
+</packages>
